requirements.txt
PyQt6
hidapi
pymodbus
